Add custom_reader for archives
It is fairly typical for Python file objects to be used on the read side
of archive processing. For instance, io.BufferedReader and similar
API that handles I/O itself.

To allow for libarchive interaction with these objects, we need a
custom_reader that can use the `readinto()` style access to the raw
archive data. `readinto()` is used as that allows us to read the data
into a `ctypes.create_string_buffer()` structure and easily cast the
pointers for the read callback. The preallocated buffer also ensures
that libarchive gets access to contiguous memory and that buffer isn't
garbage collected by Python after the callback returns.